+++ flutter_local_notifications/android/src/main/java/com/dexterous/flutterlocalnotifications/ActionBroadcastReceiver.java
@@ -3,19 +3,16 @@ package com.dexterous.flutterlocalnotifications;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.os.Bundle;
 import android.util.Log;
 
 import androidx.annotation.Keep;
 import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
 import androidx.core.app.NotificationManagerCompat;
-import androidx.core.app.RemoteInput;
 
 import com.dexterous.flutterlocalnotifications.isolate.IsolatePreferences;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -31,11 +28,7 @@ import io.flutter.view.FlutterCallbackInformation;
 public class ActionBroadcastReceiver extends BroadcastReceiver {
   public static final String ACTION_TAPPED =
       "com.dexterous.flutterlocalnotifications.ActionBroadcastReceiver.ACTION_TAPPED";
-  public static final String ACTION_ID = "actionId";
-  public static final String NOTIFICATION_ID = "notificationId";
-  public static final String INPUT_RESULT = "FlutterLocalNotificationsPluginInputResult";
   private static final String TAG = "ActionBroadcastReceiver";
-  private static final String INPUT = "input";
   @Nullable private static ActionEventSink actionEventSink;
   @Nullable private static FlutterEngine engine;
   IsolatePreferences preferences;
@@ -56,26 +49,12 @@ public class ActionBroadcastReceiver extends BroadcastReceiver {
 
     preferences = preferences == null ? new IsolatePreferences(context) : preferences;
 
-    final Map<String, Object> action = new HashMap<>();
-    final int notificationId = intent.getIntExtra(NOTIFICATION_ID, -1);
-    action.put(NOTIFICATION_ID, notificationId);
-    action.put(
-        ACTION_ID, intent.hasExtra(ACTION_ID) ? intent.getStringExtra(ACTION_ID) : "unknown");
-    action.put(
-        FlutterLocalNotificationsPlugin.PAYLOAD,
-        intent.hasExtra(FlutterLocalNotificationsPlugin.PAYLOAD)
-            ? intent.getStringExtra(FlutterLocalNotificationsPlugin.PAYLOAD)
-            : "");
-
-    Bundle remoteInput = RemoteInput.getResultsFromIntent(intent);
-    if (remoteInput != null) {
-      action.put(INPUT, remoteInput.getString(INPUT_RESULT));
-    } else {
-      action.put(INPUT, "");
-    }
+    final Map<String, Object> action =
+        FlutterLocalNotificationsPlugin.extractNotificationResponseMap(intent);
 
     if (intent.getBooleanExtra(FlutterLocalNotificationsPlugin.CANCEL_NOTIFICATION, false)) {
-      NotificationManagerCompat.from(context).cancel(notificationId);
+      NotificationManagerCompat.from(context)
+          .cancel((int) action.get(FlutterLocalNotificationsPlugin.NOTIFICATION_ID));
     }
 
     if (actionEventSink == null) {