+++ flutter_local_notifications/macos/Classes/FlutterLocalNotificationsPlugin.swift
@@ -37,6 +37,8 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
         static let filePath = "filePath"
         static let threadIdentifier = "threadIdentifier"
         static let interruptionLevel = "interruptionLevel"
+        static let actionId = "actionId"
+        static let notificationResponseType = "notificationResponseType"
     }
 
     struct ErrorMessages {
@@ -73,7 +75,7 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
     var defaultPresentAlert = false
     var defaultPresentSound = false
     var defaultPresentBadge = false
-    var launchPayload: String?
+    var launchNotificationResponseDict: [String: Any?]?
     var launchingAppFromNotification = false
 
     init(fromChannel channel: FlutterMethodChannel) {
@@ -117,25 +119,25 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
         if response.actionIdentifier == UNNotificationDefaultActionIdentifier {
             let payload = response.notification.request.content.userInfo[MethodCallArguments.payload] as? String
             if initialized {
-                handleSelectNotification(payload: payload)
+                handleSelectNotification(notificationId: Int(response.notification.request.identifier)!, payload: payload)
             } else {
-                launchPayload = payload
+                launchNotificationResponseDict = extractNotificationResponseDict(response: response)
                 launchingAppFromNotification = true
             }
 
             completionHandler()
+        } else if response.actionIdentifier == UNNotificationDismissActionIdentifier {
+            completionHandler()
         } else {
-            let text = (response as? UNTextInputNotificationResponse)?.userText ?? ""
-
-            // No isolate can be used for macOS until https://github.com/flutter/flutter/issues/65222 is resolved.
-            //
-            // Therefore, we call the regular method channel and let the macos plugin handle it appropriately.
-            handleSelectNotificationAction(payload: [
-                "notificationId": response.notification.request.identifier,
-                "actionId": response.actionIdentifier,
-                "input": text,
-                "payload": response.notification.request.content.userInfo["payload"] as? String ?? ""
-            ])
+            if initialized {
+                // No isolate can be used for macOS until https://github.com/flutter/flutter/issues/65222 is resolved.
+                //
+                // Therefore, we call the regular method channel and let the macos plugin handle it appropriately.
+                handleSelectNotificationAction(arguments: extractNotificationResponseDict(response: response))
+            } else {
+                launchNotificationResponseDict = extractNotificationResponseDict(response: response)
+                launchingAppFromNotification = true
+            }
 
             completionHandler()
         }
@@ -143,7 +145,7 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
 
     public func userNotificationCenter(_ center: NSUserNotificationCenter, didActivate notification: NSUserNotification) {
         if notification.activationType == .contentsClicked {
-            handleSelectNotification(payload: notification.userInfo![MethodCallArguments.payload] as? String)
+            handleSelectNotification(notificationId: Int(notification.identifier!)!, payload: notification.userInfo![MethodCallArguments.payload] as? String)
         }
     }
 
@@ -283,7 +285,7 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
 
     func getNotificationAppLaunchDetails(_ result: @escaping FlutterResult) {
         if #available(OSX 10.14, *) {
-            let appLaunchDetails: [String: Any?] = [MethodCallArguments.notificationLaunchedApp: launchingAppFromNotification, MethodCallArguments.payload: launchPayload]
+            let appLaunchDetails: [String: Any?] = [MethodCallArguments.notificationLaunchedApp: launchingAppFromNotification, "notificationResponse": launchNotificationResponseDict]
             result(appLaunchDetails)
         } else {
             result(nil)
@@ -622,11 +624,30 @@ public class FlutterLocalNotificationsPlugin: NSObject, FlutterPlugin, UNUserNot
         return String(arguments[MethodCallArguments.id] as! Int)
     }
 
-    func handleSelectNotification(payload: String?) {
-        channel.invokeMethod("selectNotification", arguments: payload)
+    func handleSelectNotification(notificationId: Int, payload: String?) {
+        var arguments: [String: Any?] = [:]
+        arguments["notificationId"] = notificationId
+        arguments["payload"] = payload
+        arguments["notificationResponseType"] = 0
+        channel.invokeMethod("didReceiveForegroundNotificationResponse", arguments: arguments)
     }
 
-    func handleSelectNotificationAction(payload: [String: Any]) {
-        channel.invokeMethod("selectNotificationAction", arguments: payload)
+    func handleSelectNotificationAction(arguments: [String: Any?]) {
+        channel.invokeMethod("didReceiveForegroundNotificationResponse", arguments: arguments)
+    }
+
+    @available(macOS 10.14, *)
+    func extractNotificationResponseDict(response: UNNotificationResponse) -> [String: Any?] {
+        var notificationResponseDict: [String: Any?] = [:]
+        notificationResponseDict["notificationId"] = Int(response.notification.request.identifier)!
+        if response.actionIdentifier == UNNotificationDefaultActionIdentifier {
+            notificationResponseDict[MethodCallArguments.notificationResponseType] = 0
+        } else if response.actionIdentifier != UNNotificationDismissActionIdentifier {
+            notificationResponseDict[MethodCallArguments.actionId] = response.actionIdentifier
+            notificationResponseDict[MethodCallArguments.notificationResponseType] = 1
+        }
+        notificationResponseDict["input"] = (response as? UNTextInputNotificationResponse)?.userText
+        notificationResponseDict[MethodCallArguments.payload] = response.notification.request.content.userInfo[MethodCallArguments.payload]
+        return notificationResponseDict
     }
 }