+++ flutter_local_notifications_linux/lib/src/flutter_local_notifications_platform_linux.dart
@@ -13,13 +13,13 @@ abstract class FlutterLocalNotificationsPlatformLinux
   /// Initializes the plugin.
   ///
   /// Call this method on application before using the plugin further.
-  /// 
-  /// [onSelectNotificationAction] specifies a callback handler which receives
-  /// notification action IDs.
+  ///
+  /// [onDidReceiveForegroundNotificationResponse] specifies a callback handler
+  /// which receives notification action IDs.
   Future<bool?> initialize(
     LinuxInitializationSettings initializationSettings, {
-    SelectNotificationCallback? onSelectNotification,
-    SelectNotificationActionCallback? onSelectNotificationAction,
+    DidReceiveForegroundNotificationResponseCallback?
+        onDidReceiveForegroundNotificationResponse,
   });
 
   /// Show a notification with an optional payload that will be passed back to