+++ flutter_local_notifications_linux/lib/src/flutter_local_notifications_stub.dart
@@ -23,10 +23,11 @@ class LinuxFlutterLocalNotificationsPlugin
   @override
   Future<bool?> initialize(
     LinuxInitializationSettings initializationSettings, {
-    SelectNotificationCallback? onSelectNotification,
-    SelectNotificationActionCallback? onSelectNotificationAction,
+    DidReceiveForegroundNotificationResponseCallback?
+        onDidReceiveForegroundNotificationResponse,
   }) async {
     assert(false);
+    return null;
   }
 
   /// Errors on attempted calling of the stub. It exists only to satisfy