+++ flutter_local_notifications_platform_interface/lib/src/typedefs.dart
@@ -1,9 +1,11 @@
 import 'types.dart';
 
-/// Signature of callback passed that is triggered when user selects a
-/// notification.
-typedef SelectNotificationCallback = void Function(String? payload);
+/// Signature of callback triggered when a user taps on a notification or
+/// a notification action
+typedef DidReceiveForegroundNotificationResponseCallback = void Function(
+    NotificationResponse details);
 
-/// Signature of callback triggered when a user selects on a notification action
-typedef SelectNotificationActionCallback = void Function(
-    NotificationActionDetails details);
+/// Signature of callback triggered when a user taps on a notification or
+/// a notification action
+typedef DidReceiveBackgroundNotificationResponseCallback = void Function(
+    NotificationResponse details);