+++ flutter_local_notifications_platform_interface/lib/src/types.dart
@@ -73,25 +73,56 @@ class ActiveNotification {
 }
 
 /// Details of a Notification Action that was triggered.
-class NotificationActionDetails {
-  /// Constructs an instance of [NotificationActionDetails]
-  NotificationActionDetails({
-    required this.id,
-    required this.actionId,
-    required this.input,
-    required this.payload,
+class NotificationResponse {
+  /// Constructs an instance of [NotificationResponse]
+  const NotificationResponse({
+    required this.notificationResponseType,
+    this.id,
+    this.actionId,
+    this.input,
+    this.payload,
   });
 
   /// The notification's id.
-  final int id;
+  ///
+  /// This is nullable as support for this only supported for notifications
+  /// created using version 10 or newer of this plugin.
+  final int? id;
 
   /// The id of the action that was triggered.
-  final String actionId;
+  final String? actionId;
 
   /// The value of the input field if the notification action had an input
   /// field.
   final String? input;
 
-  /// The notification's payload
+  /// The notification's payload.
   final String? payload;
+
+  /// The notification response type.
+  final NotificationResponseType notificationResponseType;
+}
+
+/// Contains details on the notification that launched the application.
+class NotificationAppLaunchDetails {
+  /// Constructs an instance of [NotificationAppLaunchDetails].
+  const NotificationAppLaunchDetails(
+    this.didNotificationLaunchApp, {
+    this.notificationResponse,
+  });
+
+  /// Indicates if the app was launched via notification.
+  final bool didNotificationLaunchApp;
+
+  /// Contains details of the notification that launched the app.
+  final NotificationResponse? notificationResponse;
+}
+
+/// The possible notification response types
+enum NotificationResponseType {
+  /// Indicates that a user has selected a notification.
+  selectedNotification,
+
+  /// Indicates the a user has selected a notification action.
+  selectedNotificationAction,
 }